# openapi/api.json
{
  "components": {
    "schemas": {
      "LolChampionsCollectionsChampionChroma": {
        "type": "object",
        "properties": {
          "chromaPath": {
            "type": "string"
          },
          "championId": {
            "format": "int32",
            "type": "integer"
          },
          "stillObtainable": {
            "type": "boolean"
          },
          "ownership": {
            "$ref": "#/components/schemas/LolChampionsCollectionsOwnership"
          },
          "lastSelected": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "colors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LolChampionsCollectionsRental": {
        "type": "object",
        "properties": {
          "purchaseDate": {
            "format": "int64",
            "type": "integer"
          },
          "rented": {
            "type": "boolean"
          },
          "endDate": {
            "format": "int64",
            "type": "integer"
          },
          "winCountRemaining": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
...
      "LolChampionsCollectionsChampionSkinEmblemPosition": {
        "type": "object",
        "properties": {
          "horizontal": {
            "type": "string"
          },
          "vertical": {
            "type": "string"
          }
        }
      },
      "LolChampionsCollectionsChampionSkinEmblemPath": {
        "type": "object",
        "properties": {
          "small": {
            "type": "string"
          },
          "large": {
            "type": "string"
          }
        }
      },
      "LolChampionsCollectionsChampionSkin": {
        "type": "object",
        "properties": {
          "rarityGemPath": {
            "type": "string"
          },
          "loadScreenPath": {
            "type": "string"
          },
          "championId": {
            "format": "int32",
            "type": "integer"
          },
          "stillObtainable": {
            "type": "boolean"
          },
          "tilePath": {
            "type": "string"
          },
          "chromas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LolChampionsCollectionsChampionChroma"
            }
          },
          "lastSelected": {
            "type": "boolean"
          },
          "splashPath": {
            "type": "string"
          },
          "uncenteredSplashPath": {
            "type": "string"
          },
          "chromaPath": {
            "type": "string"
          },
          "skinType": {
            "type": "string"
          },
          "emblems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LolChampionsCollectionsChampionSkinEmblem"
            }
          },
          "featuresText": {
            "type": "string"
          },
          "ownership": {
            "$ref": "#/components/schemas/LolChampionsCollectionsOwnership"
          },
          "name": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "isBase": {
            "type": "boolean"
          },
          "splashVideoPath": {
            "type": "string"
          }
        }
      },
      "Challenge": {
        "type": "object",
        "required": [
          "availableIds",
          "category",
          "childrenIds",
          "completedIds",
          "currentLevel",
          "currentThreshold",
          "currentValue",
          "description",
          "descriptionShort",
          "gameModes",
          "hasLeaderboard",
          "iconPath",
          "id",
          "idListType",
          "name",
          "retireTimestamp",
          "thresholds"
        ],
        "properties": {
          "previousLevel": {
            "type": "string"
          },
          "retireTimestamp": {
            "format": "date-time",
            "type": "string"
          },
          "hasLeaderboard": {
            "type": "boolean"
          },
          "playersInLevel": {
            "format": "int32",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "nextThreshold": {
            "format": "double",
            "type": "number"
          },
          "source": {
            "type": "string"
          },
          "isApex": {
            "type": "boolean"
          },
          "nextLevelIconPath": {
            "type": "string"
          },
          "capstoneGroupName": {
            "type": "string"
          },
          "levelToIconPath": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "capstoneGroupId": {
            "format": "int32",
            "type": "integer"
          },
          "availableIds": {
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          },
          "idListType": {
            "type": "string"
          },
          "valueMapping": {
            "type": "string"
          },
          "retired": {
            "type": "boolean"
          },
          "friendsAtLevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendsAtLevel"
            }
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "currentLevelAchievedTime": {
            "format": "int64",
            "type": "integer"
          },
          "childrenIds": {
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          },
          "pointsAwarded": {
            "format": "int32",
            "type": "integer"
          },
          "nextLevel": {
            "type": "string"
          },
          "descriptionShort": {
            "type": "string"
          },
          "priority": {
            "format": "double",
            "type": "number"
          },
          "previousValue": {
            "format": "double",
            "type": "number"
          },
          "parentId": {
            "format": "int32",
            "type": "integer"
          },
          "isCapstone": {
            "type": "boolean"
          },
          "currentLevel": {
            "$ref": "#/components/schemas/Leagues"
          },
          "currentThreshold": {
            "format": "double",
            "type": "number"
          },
          "gameModes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "parentName": {
            "type": "string"
          },
          "thresholds": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Reward"
            },
            "type": "object"
          },
          "percentile": {
            "format": "double",
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isReverseDirection": {
            "type": "boolean"
          },
          "completedIds": {
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          },
          "position": {
            "format": "int32",
            "type": "integer"
          },
          "category": {
            "type": "string"
          },
          "iconPath": {
            "type": "string"
          },
          "currentValue": {
            "format": "double",
            "type": "number"
          }
        }
      },
      "LolChampionsCollectionsOwnership": {
        "type": "object",
        "properties": {
          "owned": {
            "type": "boolean"
          },
          "freeToPlayReward": {
            "type": "boolean"
          },
          "rental": {
            "$ref": "#/components/schemas/LolChampionsCollectionsRental"
          }
        }
      },
      "FriendsAtLevel": {
        "type": "object",
        "properties": {
          "level": {
            "$ref": "#/components/schemas/Leagues"
          },
          "friends": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LolChampionsCollectionsChampionSkinEmblem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "positions": {
            "$ref": "#/components/schemas/LolChampionsCollectionsChampionSkinEmblemPosition"
          },
          "emblemPath": {
            "$ref": "#/components/schemas/LolChampionsCollectionsChampionSkinEmblemPath"
          }
        }
      },
      "Champion": {
        "type": "object",
        "required": [
          "id",
          "laneAssignments",
          "name",
          "roles",
          "squarePortraitJpg",
          "squarePortraitPath"
        ],
        "properties": {
          "skins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LolChampionsCollectionsChampionSkin"
            }
          },
          "baseLoadScreenPath": {
            "type": "string"
          },
          "baseSplashPath": {
            "type": "string"
          },
          "purchased": {
            "type": "integer"
          },
          "squarePortraitJpg": {
            "type": "array",
            "items": {
              "format": "byte",
              "type": "string"
            }
          },
          "laneAssignments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "squarePortraitPath": {
            "type": "string"
          }
        }
      },
...
      "Reward": {
        "type": "object",
        "properties": {
          "quantity": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "Leagues": {
        "type": "string",
        "enum": [
          "IRON",
          "BRONZE",
          "SILVER",
          "GOLD",
          "PLATINUM",
          "EMERALD",
          "DIAMOND",
          "MASTER",
          "GRANDMASTER",
          "CHALLENGER",
          "NONE"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8081"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/api/lcu-challenges/challenge": {
      "get": {
        "operationId": "getChallengeInfo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "challengeName",
            "required": true
          }
        ],
        "tags": [
          "lcu-controller"
        ]
      }
    },
...
    "/api/lcu-challenges/execute-request": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "operationId": "executeApiRequest",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "lcu-controller"
        ]
      }
    },
...
    "/api/lcu-challenges/skin": {
      "get": {
        "operationId": "getSkinForId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LolChampionsCollectionsChampionSkin"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "name": "skinId",
            "required": true
          }
        ],
        "tags": [
          "lcu-controller"
        ]
      }
    },
...
    "/api/lcu-challenges/champion": {
      "get": {
        "operationId": "getChampionForId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Champion"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "name": "championId",
            "required": true
          }
        ],
        "tags": [
          "lcu-controller"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}